requirements.txt
numpy
pillow
dataclasses
